+++ package.json
@@ -21,6 +21,6 @@
   "author": "Larry Marburger",
   "license": "MIT",
   "dependencies": {
-    "heroku-client": "^1.9.1"
+    "heroku-cli-util": "^3.5.2"
   }
 }
